Revert "Converts ListInputModal to actually be a Modal | Adds ListInputWindow…"
This reverts commit b218130a2c92e5da10d1f39f23256758ef503edc.
