+++ code/modules/tgui_input/list.dm
@@ -111,7 +111,7 @@
 /datum/tgui_list_input/ui_interact(mob/user, datum/tgui/ui)
 	ui = SStgui.try_update_ui(user, src, ui)
 	if(!ui)
-		ui = new(user, src, "ListInputWindow")
+		ui = new(user, src, "ListInputModal")
 		ui.open()
 
 /datum/tgui_list_input/ui_close(mob/user)
